+++ lib/op25_repeater/lib/dmr_bs_tx_bb_impl.cc
@@ -50,7 +50,7 @@ static void print_result(char title[], const uint8_t r[], int len) {
 }
 #endif
 
-static inline int store_i(int reg, uint8_t val[], int len) {
+static inline void store_i(int reg, uint8_t val[], int len) {
 	for (int i=0; i<len; i++){
 		val[i] = (reg >> (len-1-i)) & 1;
 	}