+++ trunk-recorder/main.cc
@@ -54,15 +54,9 @@
 #include "systems/p25_parser.h"
 #include "systems/parser.h"
 
-#include "uploaders/stat_socket.h"
 
-#include <websocketpp/config/asio_no_tls_client.hpp>
-#include <websocketpp/client.hpp>
 
-// This header pulls in the WebSocket++ abstracted thread support that will
-// select between boost::thread and std::thread based on how the build system
-// is configured.
-#include <websocketpp/common/thread.hpp>
+
 
 #include <osmosdr/source.h>
 
@@ -94,8 +88,18 @@ SmartnetParser *smartnet_parser;
 P25Parser *p25_parser;
 
 Config config;
-stat_socket stats;
+
 string default_mode;
+
+#include <websocketpp/config/asio_no_tls_client.hpp>
+#include <websocketpp/client.hpp>
+
+// This header pulls in the WebSocket++ abstracted thread support that will
+// select between boost::thread and std::thread based on how the build system
+// is configured.
+#include <websocketpp/common/thread.hpp>
+#include "uploaders/stat_socket.h"
+stat_socket stats;
 
 
 void exit_interupt(int sig) { // can be called asynchronously
@@ -294,6 +298,7 @@ void load_config(string config_file)
                   pt.get_child("sources"))
     {
       bool   qpsk_mod       = true;
+      bool   gain_set       = false;
       int    silence_frames = node.second.get<int>("silenceFrames", 0);
       double center         = node.second.get<double>("center", 0);
       double rate           = node.second.get<double>("rate", 0);
@@ -308,7 +313,6 @@ void load_config(string config_file)
       int    tia_gain       = node.second.get<double>("tiaGain", 0);
       int    vga1_gain      = node.second.get<double>("vga1Gain", 0);
       int    vga2_gain      = node.second.get<double>("vga2Gain", 0);
-      double fsk_gain       = node.second.get<double>("fskGain", 1.0);
       double digital_levels = node.second.get<double>("digitalLevels", 1.0);
       double analog_levels  = node.second.get<double>("analogLevels", 8.0);
       double squelch_db     = node.second.get<double>("squelch", 0);
@@ -378,36 +382,56 @@ void load_config(string config_file)
       BOOST_LOG_TRIVIAL(info) << "Min Freqency: " << FormatFreq(source->get_min_hz());
 
       if (if_gain != 0) {
+        gain_set = true;
         source->set_if_gain(if_gain);
       }
 
       if (bb_gain != 0) {
+        gain_set = true;
         source->set_bb_gain(bb_gain);
       }
 
       if (mix_gain != 0) {
+        gain_set = true;
         source->set_mix_gain(mix_gain);
       }
 
-      source->set_lna_gain(lna_gain);
+      if (lna_gain != 0) {
+        gain_set = true;
+        source->set_lna_gain(lna_gain);
+      }
 
-      source->set_tia_gain(tia_gain);
+      if (tia_gain != 0) {
+        gain_set = true;
+        source->set_tia_gain(tia_gain);
+      }
 
-      source->set_pga_gain(pga_gain);
-
+      if (pga_gain != 0) {
+        gain_set = true;
+        source->set_pga_gain(pga_gain);
+      }
 
       if (vga1_gain != 0) {
+        gain_set = true;
         source->set_vga1_gain(vga1_gain);
       }
 
       if (vga2_gain != 0) {
+        gain_set = true;
         source->set_vga2_gain(vga2_gain);
       }
 
-      source->set_gain(gain);
+      if (gain != 0) {
+        gain_set = true;
+        source->set_gain(gain);
+      }
+
+      if (!gain_set) {
+        BOOST_LOG_TRIVIAL(error) << "! No Gain was specified! Things will probably not work";
+      }
+
       source->set_antenna(antenna);
       source->set_squelch_db(squelch_db);
-      source->set_fsk_gain(fsk_gain);
       source->set_analog_levels(analog_levels);
       source->set_digital_levels(digital_levels);
       source->set_qpsk_mod(qpsk_mod);
@@ -640,8 +664,9 @@ void stop_inactive_recorders() {
           call->end_call();
           stats.send_call_end(call);
           call->restart_call();
-          if (recorder != NULL)
+          if (recorder != NULL) {
             stats.send_recorder(recorder);
+          }
         }
       }
       ++it;
@@ -653,8 +678,9 @@ void stop_inactive_recorders() {
         Recorder * recorder = call->get_recorder();
         call->end_call();
         stats.send_call_end(call);
-        if (recorder != NULL)
+        if (recorder != NULL) {
           stats.send_recorder(recorder);
+        }
         it = calls.erase(it);
         delete call;
       } else {
@@ -664,6 +690,7 @@ void stop_inactive_recorders() {
 
 
   }
+
 
   if (ended_recording) {
     stats.send_calls_active(calls);
@@ -717,7 +744,7 @@ bool retune_recorder(TrunkMessage message, Call *call) {
 
   if (message.freq != call->get_freq()) {
     if ((source->get_min_hz() <= message.freq) && (source->get_max_hz() >= message.freq)) {
-      recorder->tune_offset(message.freq);
+      recorder->tune_freq(message.freq);
 
       // only set the call freq, if the recorder can be retuned.
       // set the call to the new Freq / TDMA slot
@@ -863,7 +890,7 @@ void unit_check() {
     }
     myfile << "\n}\n}\n";
     sprintf(shell_command, "./unit_check.sh %s > /dev/null 2>&1 &", unit_filename);
-    system(shell_command);
+    int forget = system(shell_command);
     //int rc = system(shell_command);
     myfile.close();
   }
@@ -941,11 +968,11 @@ void retune_system(System *system) {
     if (system->get_system_type() == "smartnet") {
       // what you really need to do is go through all of the sources to find
       // the one with the right frequencies
-      system->smartnet_trunking->tune_offset(control_channel_freq);
+      system->smartnet_trunking->tune_freq(control_channel_freq);
     } else if (system->get_system_type() == "p25") {
       // what you really need to do is go through all of the sources to find
       // the one with the right frequencies
-      system->p25_trunking->tune_offset(control_channel_freq);
+      system->p25_trunking->tune_freq(control_channel_freq);
     } else {
       BOOST_LOG_TRIVIAL(error) << "\t - Unkown system type for Retune";
     }
@@ -1023,9 +1050,10 @@ void monitor_messages() {
     if (msg != 0) {
       sys_num = msg->arg1();
       sys     = find_system(sys_num);
-      sys->message_count++;
 
       if (sys) {
+        sys->message_count++;
+        
         if (sys->get_system_type() == "smartnet") {
           trunk_messages = smartnet_parser->parse_message(msg->to_string(), sys);
           handle_message(trunk_messages, sys);
@@ -1109,7 +1137,7 @@ bool monitor_system() {
             }
 
             BOOST_LOG_TRIVIAL(info) << "[" << system->get_short_name() << "]\tMonitoring Conventional Channel: " <<  FormatFreq(channel) << " Talkgroup: " << talkgroup;
-            Call_conventional *call = new Call_conventional(talkgroup, channel, system, config, &stats);
+            Call_conventional *call = new Call_conventional(talkgroup, channel, system, config);
             talkgroup++;
             Talkgroup *talkgroup = system->find_talkgroup(call->get_talkgroup());
 
@@ -1203,7 +1231,6 @@ void socket_connected()
   stats.send_config(sources, systems);
   stats.send_systems(systems);
   stats.send_calls_active(calls);
-
   std::vector<Recorder *> recorders;
 
   for (vector<Source *>::iterator it = sources.begin(); it != sources.end(); it++) {